requirements.txt
asgiref==3.11.1
Django==6.0.2
django-rest-auth==0.9.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
six==1.17.0
sqlparse==0.5.5
